ng-semantic pulled into the app and successfully mimimized and integrated.
This will allow us to keep it up to date with angular 4 deps and easily extend and improve it since it is not really maintained by anymore but by me.
I believe I can do a better job than the original author anyways. Note JQuery is still only allowed throughout the semantic code-base.